chore: 역할 분담에 따라 코드 분리
