+++ src/app/mypage/hooks/usePasswordChangeDetector.ts
@@ -3,11 +3,12 @@ import { PasswordWatchedFields } from '../types';
 export default function usePasswordChangeDetector(
   watched: PasswordWatchedFields,
 ) {
-  const isModified =
+  const isModified = Boolean(
     watched.newPassword &&
-    watched.confirmPassword &&
-    watched.newPassword === watched.confirmPassword &&
-    watched.currentPassword !== '';
+      watched.confirmPassword &&
+      watched.newPassword === watched.confirmPassword &&
+      watched.currentPassword !== '',
+  );
 
   return { isModified };
 }
